Add flyout menu open/close state
