+++ apps/frontend/src/components/footer/footer.tsx
@@ -1,0 +1,8 @@
+
+export default function Footer() {
+  return (
+    <div>
+
+    </div>
+  )
+}
